refactor: recoil 전역상태별 파일 분리
