+++ src/stores/cartItemSelected.ts
@@ -1,0 +1,8 @@
+import { atom } from "recoil";
+import persistAtom from "./utils/persist";
+
+export const isCartItemSelectedState = atom({
+  key: "isSelectedState",
+  default: {},
+  effects_UNSTABLE: [persistAtom],
+});
