+++ src/stores/cartItems.ts
@@ -1,0 +1,11 @@
+import { selector } from "recoil";
+import { getCartItems } from "../apis";
+import { CartItemType } from "../types";
+
+export const cartItemsState = selector<CartItemType[]>({
+  key: "cartItemsState",
+  get: async () => {
+    const cartItems = await getCartItems();
+    return cartItems;
+  },
+});
