+++ src/stores/utils/persist.ts
@@ -1,0 +1,8 @@
+import { recoilPersist } from "recoil-persist";
+
+const { persistAtom } = recoilPersist({
+  key: "recoil-persist",
+  storage: localStorage,
+});
+
+export default persistAtom;
